# Identify comp overhead & Sin estimate