--- a/basalt/autograd/graph.mojo
+++ b/basalt/autograd/graph.mojo
@@ -34,7 +34,9 @@ struct Graph:
         self.symbol_count += 1
         return inp
 
-    fn param(inout self, shape: TensorShape, init: Param, trainable: Bool = True) -> Symbol:
+    fn param(
+        inout self, shape: TensorShape, init: Param, trainable: Bool = True
+    ) -> Symbol:
         var param_id = Symbol(self.symbol_count, dtype, shape, trainable)
         self.params.put(param_id, init)
         self.symbol_count += 1
@@ -48,12 +50,16 @@ struct Graph:
 
     fn scalar(inout self, value: SIMD[dtype, 1]) -> Symbol:
         var scal = Param(value)
-        var scalar_id = Symbol(self.symbol_count, dtype, TensorShape(1), trainable=False)
+        var scalar_id = Symbol(
+            self.symbol_count, dtype, TensorShape(1), trainable=False
+        )
         self.params.put(scalar_id, scal)
         self.symbol_count += 1
         return scalar_id
 
-    fn constant(inout self, shape: TensorShape, data: DynamicVector[SIMD[dtype, 1]]) -> Symbol:
+    fn constant(
+        inout self, shape: TensorShape, data: DynamicVector[SIMD[dtype, 1]]
+    ) -> Symbol:
         var cst = Param(data)
         var constant_id = Symbol(self.symbol_count, dtype, shape, trainable=False)
         self.params.put(constant_id, cst)
@@ -65,27 +71,36 @@ struct Graph:
 
     fn loss(inout self, symbol: Symbol):
         self.loss_out = symbol
-    
-    fn op(inout self, op: OP,
+
+    fn op(
+        inout self,
+        op: OP,
         operand_1: Symbol,
         operand_2: Optional[Symbol] = None,
         operand_3: Optional[Symbol] = None,
-        attributes: AttributeVector = AttributeVector()
+        attributes: AttributeVector = AttributeVector(),
     ) -> Symbol:
-        
         var res: Symbol
         if operand_3:
             res = Symbol(
                 self.symbol_count,
                 dtype,
-                static_result_shape(op, operand_1.shape, operand_2.value().shape, operand_3.value().shape, attributes),
+                static_result_shape(
+                    op,
+                    operand_1.shape,
+                    operand_2.value().shape,
+                    operand_3.value().shape,
+                    attributes,
+                ),
                 self.result_trainable(operand_1, operand_2.value(), operand_3.value()),
             )
         elif operand_2:
             res = Symbol(
                 self.symbol_count,
                 dtype,
-                static_result_shape(op, operand_1.shape, operand_2.value().shape, attributes),
+                static_result_shape(
+                    op, operand_1.shape, operand_2.value().shape, attributes
+                ),
                 self.result_trainable(operand_1, operand_2.value()),
             )
         else:
@@ -96,43 +111,55 @@ struct Graph:
                 self.result_trainable(operand_1),
             )
 
-        self.nodes.push_back(Node(op, res, operand_1, operand_2.take(), operand_3.take(), attributes))
+        self.nodes.push_back(
+            Node(op, res, operand_1, operand_2.take(), operand_3.take(), attributes)
+        )
         self.symbol_count += 1
         return res
 
-    fn op(inout self, op: OP,
+    fn op(
+        inout self,
+        op: OP,
         operand_1: Symbol,
         operand_2: FloatLiteral,
-        attributes: AttributeVector = AttributeVector()
+        attributes: AttributeVector = AttributeVector(),
     ) -> Symbol:
-        
         var operand_2_symbol = self.scalar(operand_2)
         var res = Symbol(
             self.symbol_count,
             dtype,
-            static_result_shape(op, operand_1.shape, operand_2_symbol.shape, attributes),
+            static_result_shape(
+                op, operand_1.shape, operand_2_symbol.shape, attributes
+            ),
             self.result_trainable(operand_1),
         )
 
-        self.nodes.push_back(Node(op, res, operand_1, operand_2_symbol, attributes=attributes))
+        self.nodes.push_back(
+            Node(op, res, operand_1, operand_2_symbol, attributes=attributes)
+        )
         self.symbol_count += 1
         return res
 
-    fn op(inout self, op: OP,
+    fn op(
+        inout self,
+        op: OP,
         operand_1: FloatLiteral,
         operand_2: Symbol,
-        attributes: AttributeVector = AttributeVector()
+        attributes: AttributeVector = AttributeVector(),
     ) -> Symbol:
-        
         var operand_1_symbol = self.scalar(operand_1)
         var res = Symbol(
             self.symbol_count,
             dtype,
-            static_result_shape(op, operand_1_symbol.shape, operand_2.shape, attributes),
+            static_result_shape(
+                op, operand_1_symbol.shape, operand_2.shape, attributes
+            ),
             self.result_trainable(operand_2),
         )
 
-        self.nodes.push_back(Node(op, res, operand_1_symbol, operand_2, attributes=attributes))
+        self.nodes.push_back(
+            Node(op, res, operand_1_symbol, operand_2, attributes=attributes)
+        )
         self.symbol_count += 1
         return res
 
@@ -145,7 +172,9 @@ struct Graph:
         return operand_1.trainable or operand_2.trainable
 
     @staticmethod
-    fn result_trainable(operand_1: Symbol, operand_2: Symbol, operand_3: Symbol) -> Bool:
+    fn result_trainable(
+        operand_1: Symbol, operand_2: Symbol, operand_3: Symbol
+    ) -> Bool:
         return operand_1.trainable or operand_2.trainable or operand_3.trainable
 
     fn json(self) -> String: