--- a/basalt/utils/string_dict/__init__.mojo
+++ b/basalt/utils/string_dict/__init__.mojo
@@ -1,2 +1,0 @@
-# This code is based on https://github.com/mzaks/compact-dict/tree/main
-from .dict import StringDict