This PR tries to identify and fix why there is so much overhead when using certain datatypes as compile time constants. And adds a new benchmark, `sin_estimate.mojo` to evaluate the result. 

Example: Removing UUID from Symbol results in an increase in performance.
```
Before:
housing.mojo:  2.8 sec
Time using custom Tensor & references on (AnyPointer) collection:  14.337211644 s
Time using custom Tensor & __get_address_as_lvalue on (AnyPointer) collection:  9.5543818009999999 s

After replacing UUID with a UInt32:
housing.mojo: 0.08 sec
Time using custom Tensor & references on (AnyPointer) collection:  0.243069382 s
Time using custom Tensor & __get_address_as_lvalue on (AnyPointer) collection:  0.16846102700000001 s
```

**BUG**: sin_estimate.mojo seem to be stagnating at a loss of mse 0.06. For 10000 iterations
```
pytorch: 4.15 sec, mse: 1.386e-07
basalt: 12.9 sec, mse: 0.0603
```