@StijnWoestenborghs 2024-03-24
Removing UUID entirely here. The root cause was not uuid on itself being slow in generating. But the String type being wildly inefficient. UUID has been replaced in symbol by a simple counter in the graph of type UInt32. And with uuid being removed there is no need for a StringDict inside the Collection. Reducing overhead and complexity and increased the performance to:
```
housing.mojo: 0.04 sec
Time using custom Tensor & references on (AnyPointer) collection:  0.010930570000000001 s
Time using custom Tensor & __get_address_as_lvalue on (AnyPointer) collection:  0.0095764560000000006 s
```