@StijnWoestenborghs 2024-03-25
All changes:
- Removes UUID and StringDict. Symbol.name replaced with a UInt32. Collection changed accordingly.
- Adds a profiler to the main branch (I think I don't want to add this permanently though, but I guess for now it can be there temporarily, as it is pretty convenient)
- Adds a new bench mark: sin_estimate.mojo
- Fixes a bug in vectorization of the ReLU operator.
- Optimizations on the `get_real_index` function used for broadcasting and accumulating gradients.
New performance on the benchmark (double the amount of epochs with previous report):
```
sin_estimate.mojo
pytorch: 7.73 sec, mse: 1.0341e-05
basalt: 1.243 sec, mse: 2.7471e-05
```