@StijnWoestenborghs 2024-03-25
Getting the bugs out but we are starting to see the benefits of the static graph 🚀
Can't wait for matmul/conv